requirements.txt
python-dotenv
openai
streamlit
rank_bm25
langchain_openai
langchain_chroma